Refine historical context in Unix origins post
https://github.com/joone/fosscomics/issues/3

Update the post on the origins of Unix and the C language to clarify
the distinct roles and contributions in the development of ITS and Unix.
Specifically, delineate the involvement of MIT folks in ITS and
highlight the contributions of Ken Thompson and Dennis Ritchie
from Bell Labs in creating Unix.